requirements.txt
bleak
paho-mqtt
boto3
aws-iot-device-sdk
